% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_payments.R
\name{billys_bank_payments}
\alias{billys_bank_payments}
\title{Function to get the data from the bank account.}
\usage{
billys_bank_payments(...)
}
\description{
Function to get the data from the bank account.
}
